Create Hot_Potato_Game
Hot potato is a game in which children form a circle and start passing a hot potato. The counting starts with the first kid. Every nth toss the child, which is left with the potato, leaves the game. Consequently, it passes the potato forward until only one kid left.

We print every kid that is removed from the circle, and in the end we print the last child.